+++ tardis/montecarlo/montecarlo_numba/single_packet_loop.py
@@ -31,8 +31,8 @@ from tardis import constants as const
 
 C_SPEED_OF_LIGHT = const.c.to("cm/s").value
 
-from tardis.io.logger.montecarlo_logger import log_decorator
-from tardis.io.logger import montecarlo_logger as mc_logger
+from tardis.io.logger.montecarlo_tracking import log_decorator
+from tardis.io.logger import montecarlo_tracking as mc_tracker
 
 # @log_decorator
 @njit
@@ -67,7 +67,7 @@ def single_packet_loop(
         r_packet, vpacket_collection, numba_model, numba_plasma
     )
 
-    if mc_logger.DEBUG_MODE:
+    if mc_tracker.DEBUG_MODE:
         r_packet_track_nu = [r_packet.nu]
         r_packet_track_mu = [r_packet.mu]
         r_packet_track_r = [r_packet.r]
@@ -114,14 +114,14 @@ def single_packet_loop(
             trace_vpacket_volley(
                 r_packet, vpacket_collection, numba_model, numba_plasma
             )
-        if mc_logger.DEBUG_MODE:
+        if mc_tracker.DEBUG_MODE:
             r_packet_track_nu.append(r_packet.nu)
             r_packet_track_mu.append(r_packet.mu)
             r_packet_track_r.append(r_packet.r)
             r_packet_track_interaction.append(interaction_type)
             r_packet_track_distance.append(distance)
 
-    if mc_logger.DEBUG_MODE:
+    if mc_tracker.DEBUG_MODE:
         return (
             r_packet_track_nu,
             r_packet_track_mu,
